+++ c/driver/sqlite/sqlite.cc
@@ -26,11 +26,11 @@
 #define ADBC_FRAMEWORK_USE_FMT
 #include "driver/common/options.h"
 #include "driver/common/utils.h"
-#include "driver/framework/base_connection.h"
-#include "driver/framework/base_database.h"
 #include "driver/framework/base_driver.h"
-#include "driver/framework/base_statement.h"
 #include "driver/framework/catalog.h"
+#include "driver/framework/connection.h"
+#include "driver/framework/database.h"
+#include "driver/framework/statement.h"
 #include "driver/framework/status.h"
 #include "driver/sqlite/statement_reader.h"
